+++ e2e/tests/text.e2e.ts
@@ -260,7 +260,7 @@ describe('text', () => {
 
   it('should not allow applying link to atomic blocks in selection', function () {
     cy.loadRicosEditorAndViewer('content-with-video');
-    cy.waitForVideoToLoad();
+    cy.waitForMediaToLoad();
     cy.setEditorSelection(0, 5);
     cy.getInlineButton(INLINE_TOOLBAR_BUTTONS.LINK).should('not.be.disabled');
     cy.eyesCheckWindow(this.test.title);