+++ packages/plugin-audio/web/src/toolbar/inline-buttons.tsx
@@ -41,23 +41,6 @@ const createInlineButtons: CreateInlineButtons = ({
     : undefined;
 
   return [
-    {
-      keyName: 'settings',
-      type: BUTTONS.AUDIO_SETTINGS,
-      fullHeight: true,
-      icon: PluginSettingsIcon,
-      modalName: Modals.AUDIO_SETTINGS,
-      modalStyles: getModalStyles({
-        isMobile,
-      }),
-      t,
-      modalElement: AudioSettings,
-      mobile: true,
-      tooltipTextKey: 'SettingsButton_Tooltip',
-      settings,
-      triggerSettingsBi: true,
-      pluginId: AUDIO_TYPE,
-    },
     { keyName: 'sizeSmallCenter', type: BUTTONS.SIZE_SMALL_CENTER, mobile: false },
     { keyName: 'sizeContent', type: BUTTONS.SIZE_CONTENT, mobile: false },
     { keyName: 'sizeSmallLeft', type: BUTTONS.SIZE_SMALL_LEFT, mobile: false },
@@ -90,6 +73,23 @@ const createInlineButtons: CreateInlineButtons = ({
           },
         };
       },
+    },
+    {
+      keyName: 'settings',
+      type: BUTTONS.AUDIO_SETTINGS,
+      fullHeight: true,
+      icon: PluginSettingsIcon,
+      modalName: Modals.AUDIO_SETTINGS,
+      modalStyles: getModalStyles({
+        isMobile,
+      }),
+      t,
+      modalElement: AudioSettings,
+      mobile: true,
+      tooltipTextKey: 'SettingsButton_Tooltip',
+      settings,
+      triggerSettingsBi: true,
+      pluginId: AUDIO_TYPE,
     },
     { keyName: 'delete', type: BUTTONS.DELETE, mobile: true },
   ];